fix(add-money): choose deposit method once, not twice
The add-money flow asked "Bank or Crypto" twice — once on a standalone
method screen, then again on the per-country screen after picking a
country. Customers reported this as buggy/confusing (TASK-20033).

Collapse the redundant first screen: /add-money now lands directly on the
country list (with a crypto shortcut + the offramp-migrate entry), and the
country-aware per-country screen remains the single place the method is
chosen. The per-country screen already knows each country's real methods
and coming-soon states, so no country regresses.
