+++ src/app/(mobile-ui)/add-money/page.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import AddMoneyMethodSelection from '@/components/AddMoney/views/AddMoneyMethodSelection.view'
 import AddWithdrawCountriesList from '@/components/AddWithdraw/AddWithdrawCountriesList'
 import dynamic from 'next/dynamic'
 
@@ -9,12 +8,19 @@ const OnrampBankPage = dynamic(() => import('./_onramp-bank'), { ssr: false })
 const OnrampMantecaPage = dynamic(() => import('./_onramp-manteca'), { ssr: false })
 import { CountryList } from '@/components/Common/CountryList'
 import type { CountryData } from '@/components/AddMoney/consts'
+import { ActionListCard } from '@/components/ActionListCard'
+import AvatarWithBadge from '@/components/Profile/AvatarWithBadge'
+import ChooseNetworkDrawer from '@/components/AddMoney/components/ChooseNetworkDrawer'
+// offramp.xyz migrants get this link-granted badge at signup (peanut-api-ts
+// invite/badge routes, code `offramp` / utm `offramp`).
+import { OFFRAMP_BADGE_CODE } from '@/components/Invites/campaign-maps'
+import type { RhinoChainType } from '@/services/services.types'
 import NavHeader from '@/components/Global/NavHeader'
+import { useAuth } from '@/context/authContext'
 import { useOnrampFlow } from '@/context/OnrampFlowContext'
 import { useRouter, useSearchParams } from 'next/navigation'
-import { useEffect } from 'react'
-import { useQueryState, parseAsStringEnum } from 'nuqs'
-import { checkIfInternalNavigation, getRedirectUrl, clearRedirectUrl, getFromLocalStorage } from '@/utils/general.utils'
+import { useEffect, useState } from 'react'
+import { getRedirectUrl, clearRedirectUrl, getFromLocalStorage } from '@/utils/general.utils'
 import posthog from 'posthog-js'
 import { ANALYTICS_EVENTS } from '@/constants/analytics.consts'
 import { addMoneyCountryUrl } from '@/utils/native-routes'
@@ -23,25 +29,23 @@ export default function AddMoneyPage() {
     const router = useRouter()
     const searchParams = useSearchParams()
     const { resetOnrampFlow } = useOnrampFlow()
-    const [method, setMethod] = useQueryState('method', parseAsStringEnum(['bank']))
+    const { user } = useAuth()
+    const [isDrawerOpen, setIsDrawerOpen] = useState(false)
 
     // native app passes country as query param instead of path segment
     const countryFromQuery = searchParams.get('country')
+
+    // offramp migrants get a tailored arbitrum deposit entry above the country list
+    const hasOfframpBadge = user?.user?.badges?.some((b) => b.code === OFFRAMP_BADGE_CODE) ?? false
 
     useEffect(() => {
         if (!countryFromQuery) resetOnrampFlow()
     }, [])
 
     const handleBack = () => {
-        // if viewing country-specific form, go back to country list
+        // native country sub-view → back to the country list root
         if (countryFromQuery) {
-            router.push('/add-money?method=bank')
-            return
-        }
-
-        // if on country list view, go back to method selection
-        if (method === 'bank') {
-            setMethod(null)
+            router.push('/add-money')
             return
         }
 
@@ -71,6 +75,11 @@ export default function AddMoneyPage() {
         router.push(addMoneyCountryUrl(country.path))
     }
 
+    const handleNetworkSelect = (network: RhinoChainType) => {
+        setIsDrawerOpen(false)
+        router.push(`/add-money/crypto?network=${network}`)
+    }
+
     // native app: render sub-views based on query params
     const viewFromQuery = searchParams.get('view')
     if (countryFromQuery && viewFromQuery === 'bank') {
@@ -88,16 +97,29 @@ export default function AddMoneyPage() {
         <div className="flex min-h-[inherit] flex-col gap-8">
             <NavHeader title="Add Money" onPrev={handleBack} />
 
-            {method === 'bank' ? (
-                <CountryList
-                    inputTitle="Select your country"
-                    viewMode="add-withdraw"
-                    flow="add"
-                    onCountryClick={handleCountryClick}
+            {hasOfframpBadge && (
+                <ActionListCard
+                    title="Migrate from Offramp"
+                    description="Move your Offramp balance to Peanut"
+                    position="single"
+                    leftIcon={<AvatarWithBadge icon="wallet-outline" size="extra-small" className="bg-yellow-1" />}
+                    onClick={() => router.push('/add-money/crypto?network=EVM&source=offramp')}
                 />
-            ) : (
-                <AddMoneyMethodSelection onBankTransferClick={() => setMethod('bank')} />
             )}
+
+            <CountryList
+                inputTitle="Select your country"
+                viewMode="add-withdraw"
+                flow="add"
+                onCountryClick={handleCountryClick}
+                onCryptoClick={() => setIsDrawerOpen(true)}
+            />
+
+            <ChooseNetworkDrawer
+                open={isDrawerOpen}
+                onClose={() => setIsDrawerOpen(false)}
+                onSelect={handleNetworkSelect}
+            />
         </div>
     )
 }
